+++ gluecodium/src/main/java/com/here/gluecodium/validator/LimeStructsValidator.kt
@@ -95,7 +95,7 @@ internal class LimeStructsValidator(private val logger: LimeLogger, private val 
             logger.error(limeStruct, "an immutable struct should have at least one explicit constructor")
             result = false
         }
-        if (limeStruct.internalFields.any { it.defaultValue == null }) {
+        if (!limeStruct.visibility.isInternal && limeStruct.internalFields.any { it.defaultValue == null }) {
             logger.error(
                 limeStruct,
                 "if any internal field does not have a default value, " +